Remove index request, move all index login to server.
