--- api/src/main/java/com/alibaba/nacos/api/ai/remote/request/IndexMcpServerRequest.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 1999-2025 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.api.ai.remote.request;
-
-/**
- * Nacos AI module index mcp name to mcp Id request.
- *
- * @author xiweng.yy
- */
-public class IndexMcpServerRequest extends AbstractMcpRequest {
-    
-    private String mcpName;
-    
-    public String getMcpName() {
-        return mcpName;
-    }
-    
-    public void setMcpName(String mcpName) {
-        this.mcpName = mcpName;
-    }
-}
