+++ api/src/main/resources/META-INF/services/com.alibaba.nacos.api.remote.Payload
@@ -75,8 +75,6 @@ com.alibaba.nacos.api.naming.remote.response.NamingFuzzyWatchSyncResponse
 com.alibaba.nacos.api.lock.remote.request.LockOperationRequest
 com.alibaba.nacos.api.lock.remote.response.LockOperationResponse
 
-com.alibaba.nacos.api.ai.remote.request.IndexMcpServerRequest
-com.alibaba.nacos.api.ai.remote.response.IndexMcpServerResponse
 com.alibaba.nacos.api.ai.remote.request.QueryMcpServerRequest
 com.alibaba.nacos.api.ai.remote.response.QueryMcpServerResponse
 com.alibaba.nacos.api.ai.remote.request.ReleaseMcpServerRequest
